pyproject.toml
[project]
name = "finagent-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.5",
    "anthropic>=0.84.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.133.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "yfinance>=1.2.0",
]
